requirements.txt
certifi==2025.4.26
charset-normalizer==3.4.1
dotenv==0.9.9
idna==3.10
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
requests==2.32.3
six==1.17.0
urllib3==2.4.0
